requirements.txt
discum @ git+https://github.com/Merubokkusu/Discord-S.C.U.M.git@c8d587b7cbb59eccea00ddacd5f494b8597128f4
pandas==1.3.1
python-decouple==3.4

